--- src/components/movie/Movies.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class Movies extends Component {
-  render() {
-    return (
-      <div>
-        <p>I am the movie search</p>    
-      </div>
-    );
-  }
-}
-
-export default Movies;
