+++ src/components/reviews/Reviews.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Reviews extends Component {
+  render() {
+    return (
+      <div>
+        <p>I am the reviews</p>
+      </div>
+    );
+  }
+}
+
+export default Reviews;
